pyproject.toml
[project]
name = "sauna-hack"
version = "0.1.0"
description = "Harvia Sauna API Demo - exercises all API endpoints"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0"
]
